Was refator the auth module to include a router with the login routes and the controlers folder in webapp was deleted to just keep  one controller file, and rename models folder to database
