+++ app/auth/routes.js
@@ -1,0 +1,13 @@
+'use strict'
+const express = require('express');
+const passport = require('./passport');
+const router = express.Router();
+
+router.get('/twitter', passport.authenticate('twitter'));
+router.get('/twitter/callback', passport.authenticate('twitter', {
+	  successRedirect: '/welcome',
+    failureRedirect: '/login' 
+}));
+
+
+module.exports = router;
