+++ app/auth/twitter.js
@@ -1,10 +1,10 @@
 'use strict'
 const Twitter = require('passport-twitter').Strategy
-const User = require('../models/user')
+const User = require('../database/user')
 
 module.exports = new Twitter({
-	consumerKey		 : TWITTER_CONSUMMER_KEY,
-	consumerSecret : TWITTER_CONSUMMER_sECRET,
+	consumerKey		 : 'TWITTER_CONSUMMER_KEY',
+	consumerSecret : 'TWITTER_CONSUMMER_sECRET',
 	callbackURL		 : 'http://127.0.0.1:3000/login/twitter/callback'
 }, (token, tokenSecret, profile, done) => {
 	User.fetchOne({
