+++ app/webapp/routes.js
@@ -1,19 +1,13 @@
 'use strict'
 const express = require('express');
-const homeController = require('./controllers/home');
-const auth = require('../auth');
+const homeController = require('./controller');
 const authMiddleware = require('../middlewares/auth');
 const router = express.Router();
 
+router.get('', homeController.home);
 router.get('/login', homeController.login);
-router.get('/login/twitter', auth.authenticate('twitter'));
-router.get('/login/twitter/callback', auth.authenticate('twitter', {
-	  successRedirect: '/welcome',
-    failureRedirect: '/login' 
-}));
 router.get('/logout', homeController.logout);
 router.get('/welcome', authMiddleware.ensure, homeController.welcome);
-router.get('', homeController.home);
 
 
 module.exports = router;
